% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNS_EM.R
\name{CBR_EM}
\alias{CBR_EM}
\title{CBR_EM function:}
\usage{
CBR_EM(Z, X, ID)
}
\arguments{
\item{Z}{123}

\item{X}{123}

\item{ID}{123}
}
\value{
An list involves coefficients and fitted values.
}
\description{
description here
}
\examples{
testData <- DGP(50, 3, runif(sample(5:21, 1), -5, 5))
a <- milr(testData$Z, testData$X, testData$ID)
coef(a)
}

