% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{DGP}
\alias{DGP}
\title{DGP: data generation}
\usage{
DGP(n, m, beta)
}
\arguments{
\item{n}{The number of observations.}

\item{m}{The size of bag for each observation.}

\item{beta}{The coefficients.}
}
\value{
An list includes
}
\description{
description here
}
\examples{
data1 <- DGP(50, 3, runif(10, -5, 5))
data2 <- DGP(50, sample(3:5, 50, TRUE), runif(10, -5, 5))
}

