% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milr.R
\name{fitted.milr}
\alias{fitted.milr}
\title{Fitted Response of milr Fits}
\usage{
\method{fitted}{milr}(object, type = "bag", ...)
}
\arguments{
\item{object}{A fitted obejct of class inheriting from \code{"milr"}.}

\item{type}{The type of fitted response required. Default is \code{"bag"}, the fitted labels of bags.
The \code{"instance"} option returns the fitted labels of instances.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Fitted Response of milr Fits
}

