% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{fitted.softmax}
\alias{fitted.softmax}
\title{Fitted Response of softmax Fits}
\usage{
\method{fitted}{softmax}(object, type = "bag", ...)
}
\arguments{
\item{object}{A fitted obejct of class inheriting from \code{"softmax"}.}

\item{type}{The type of fitted response required. Default is \code{"bag"}, the fitted labels of bags.
The \code{"instance"} option returns the fitted labels of instances.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Fitted Response of softmax Fits
}

