% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milr-package.R
\docType{package}
\name{milr-package}
\alias{milr-package}
\title{The milr package: an implementation for multiple-instance logistic regression with lasso penalty}
\description{
This package performs maximum likelihood estimation for multiple-instance logistic regression  
utilizing EM algorithm. Also, LASSO penalty is implemented for variable selection.
}
\references{
\enumerate{
  \item Multiple-Instance Logistic Regression with LASSO Penalty, 
   Ray-Bing Chen, Kuang-Hung Cheng, Sheng-Mao Chang, Shuen-Lin Jeng, 
   Ping-Yang Chen, Chun-Hao Yang, and Chi-Chun Hsia, 2016
}
}

