% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milr.R
\docType{package}
\name{milr}
\alias{milr}
\alias{milr-package}
\title{The milr package: a implementation for multiple-instance logistic regression with lasso penalty}
\description{
Description here
}
\section{Reference}{

\enumerate{
  \item to be filled in.
}
}

