% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-MUSK1.R
\docType{data}
\name{MUSK1}
\alias{MUSK1}
\title{example data: MUSK1}
\format{A data.frame with 476 observations on 168 variables.
\itemize{
  \item \code{bag}: bag variable
  \item \code{label}:  label variable
  \item \code{x1 ~ x166}: coveriates
  ... 
}}
\source{
whereItComesFrom
}
\usage{
MUSK1
}
\description{
example data: MUSK1
}

