% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milr.R
\name{predict.milr}
\alias{predict.milr}
\title{Predict Method for milr Fits}
\usage{
\method{predict}{milr}(object, newdata, bag_newdata, type = "bag", ...)
}
\arguments{
\item{object}{A fitted obejct of class inheriting from \code{"milr"}.}

\item{newdata}{A matrix with variables to predict.}

\item{bag_newdata}{A vector. The labels of instances to bags.}

\item{type}{The type of prediction required. Default is \code{"bag"}, the predicted labels of bags.
The \code{"instance"} option returns the predicted labels of instances.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Predict Method for milr Fits
}

