% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{DGP}
\alias{DGP}
\title{milr function:}
\usage{
DGP(n, m, beta)
}
\arguments{
\item{n}{123}

\item{m}{123}

\item{beta}{123}
}
\value{
An list involves coefficients and fitted values.
}
\description{
description here
}
\examples{
data <- DGP(50, 3, runif(sample(5:21, 1), -5, 5))
}

