% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{DGP}
\alias{DGP}
\title{DGP: data generation}
\usage{
DGP(n, m, beta)
}
\arguments{
\item{n}{An integer. The number of subjects (bags).}

\item{m}{An integer or vector. When \code{m} is integer, subjects have the same number of instance. 
When \code{m} is vector, subjects have corresponding number of instance.}

\item{beta}{A vector. The true regression coefficients.}
}
\value{
An list includes status of subjects (bags) \code{Z}, design matrix \code{X}, subject id for all components (instances) \code{ID}.
}
\description{
Simulate multiple-instance regression data.
}
\examples{
data1 <- DGP(50, 3, runif(10, -5, 5))
data2 <- DGP(50, sample(3:5, 50, TRUE), runif(10, -5, 5))
}

