% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milr.R
\name{milr}
\alias{milr}
\title{Maximum likelihood estimation of multiple-instance logistic regression with LASSO penalty}
\usage{
milr(y, x, bag, lambda = 0, maxit = 500)
}
\arguments{
\item{y}{A vector. Binay response.}

\item{x}{The design matrix. The number of rows of x must be equal to the length of y.}

\item{bag}{A vector, bag id.}

\item{lambda}{The panelty for lasso. Default is exp(seq(log(0.01), log(50),length = 30)).
The panelty is chosen by BIC.}

\item{maxit}{An integer, the maximum iteration for EM algorithm.}
}
\value{
An list includes BIC, chosen lambda, coefficients and fitted values.
}
\description{
Please refer to \link{milr-package}.
}
\examples{
set.seed(100)
beta <- runif(5, -5, 5)
trainData <- DGP(70, 3, beta)
testData <- DGP(30, 3, beta)
milr_result <- milr(trainData$Z, trainData$X, trainData$ID,
  exp(seq(log(0.01), log(50),length = 30)))
coef(milr_result)      # coefficients
fitted(milr_result)    # fitted values
summary(milr_result)   # summary milr
predict(milr_result, testData$X, testData$ID) # predicted label

# use auto-tuning
milr_result <- milr(trainData$Z, trainData$X, trainData$ID, lambda = -1)
coef(milr_result)      # coefficients
fitted(milr_result)    # fitted values
summary(milr_result)   # summary milr
predict(milr_result, testData$X, testData$ID) # predicted label
}

