% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milr.R, R/milr_main.R
\docType{package}
\name{milr}
\alias{milr}
\alias{milr-package}
\title{The milr package: a implementation for multiple-instance logistic regression with lasso penalty}
\usage{
milr(y, x, bag = NULL, lambda = 0, alpha = 1, maxit = 500)
}
\arguments{
\item{y}{123}

\item{x}{123}

\item{bag}{123}

\item{lambda}{123}

\item{alpha}{123}

\item{maxit}{An integer, the maximum iteration for EM algorithm.}
}
\value{
An list involves coefficients and fitted values.
}
\description{
Description here

description here
}
\section{Reference}{

\enumerate{
  \item to be filled in.
}
}
\examples{
testData <- DGP(50, 3, runif(sample(5:21, 1), -5, 5))
a <- milr(testData$Z, testData$X, testData$ID)
coef(a)
}

