% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milr.R, R/milr_main.R
\docType{package}
\name{milr}
\alias{milr}
\alias{milr-package}
\title{The milr package: a implementation for multiple-instance logistic regression with lasso penalty}
\usage{
milr(y, x, bag, lambda = 0, alpha = 1, maxit = 500)
}
\arguments{
\item{y}{A vector, its elements contain two different values.}

\item{x}{The design matrix. The number of rows of x must be equal to the length of y.}

\item{bag}{bag variable.}

\item{lambda}{The panelty for lasso. Default is exp(seq(log(0.01), log(50),length = 30)).
The panelty is chosen by BIC.}

\item{alpha}{123}

\item{maxit}{An integer, the maximum iteration for EM algorithm.}
}
\value{
An list includes BIC, chosen lambda, coefficients and fitted values.
}
\description{
Description here

description here
}
\section{Reference}{

\enumerate{
  \item to be filled in.
}
}
\examples{
set.seed(100)
beta <- runif(10, -5, 5)
trainData <- DGP(70, 3, beta)
testData <- DGP(30, 3, beta)
milr_result <- milr(trainData$Z, trainData$X, trainData$ID, 
  exp(seq(log(0.01), log(50),length = 30)))
coef(milr_result)      # coefficients
fitted(milr_result)    # fitted values
predict(milr_result, testData$X, testData$ID) # predicted label
}

